chore: support decimal

+++ api/requests/food_request.go
@@ -6,11 +6,11 @@ import (
 
 type FoodRequest struct {
 	Name  string               `form:"name" binding:"required"`
-	Price int                  `form:"price" binding:"required"`
+	Price float64              `form:"price" binding:"required"`
 	Image multipart.FileHeader `form:"image" binding:"required"`
 }
 
 type UpdateFoodInput struct {
-	Name  string `json:"name"`
-	Price int    `json:"price"`
+	Name  string  `json:"name"`
+	Price float64 `json:"price"`
 }
